pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "demo-mcp-server"
version = "1.0.0"
description = "Un serveur MCP simple pour GitHub Copilot"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "fastmcp>=2.12.2",
    "websockets>=15.0.1",
]

[project.scripts]
demo-mcp-server = "server:main"

[tool.hatch.build.targets.wheel]
packages = ["."]
